Remove pluggable TaskRunners concept
As part of the changes in AIP-72 to have "activities" (i.e.
TaskExecuteActivity) and the need to shortly move the StandardTaskRunner into
the TaskSDK, and before doing that I want to simplify the strucutre.

To that end, the CGroupTaskRunner is being removed. While it (likely/probably)
still works, it's use is very uncommon as it does not work within Docker or
Kubernetes deployment (as it is not currently possible to nest cgroups inside
a container in practice).

In this PR I have removed the config option and task_runner loader concept,
deleted the CGroupRunner and merged Base and Starndard TaskRunner classes.
