+++ tests/task/conftest.py
@@ -1,4 +1,3 @@
-#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -15,3 +14,27 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+
+from __future__ import annotations
+
+import logging
+from logging.config import dictConfig
+
+import pytest
+
+
+@pytest.fixture(scope="module", autouse=True)
+def reset_to_default_logging():
+    """
+    Initialize ``BaseTaskRunner`` might have side effect to another tests.
+    This fixture reset back logging to default after execution of separate module  in this test package.
+    """
+    yield
+
+    from airflow.config_templates.airflow_local_settings import DEFAULT_LOGGING_CONFIG
+    from airflow.listeners.listener import get_listener_manager
+
+    airflow_logger = logging.getLogger("airflow")
+    airflow_logger.handlers = []
+    dictConfig(DEFAULT_LOGGING_CONFIG)
+    get_listener_manager().clear()
